# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /datasets/{id}/terrain:
    get:
      operationId: getDatasetsIdTerrain
      tags: [datasets]
      summary: Get gridded terrain data for a dataset
      description: Returns a 256×256 depth grid for the given dataset ID, suitable for building a Three.js terrain mesh
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: resolution
          required: false
          schema:
            type: integer
            default: 256
            minimum: 32
            maximum: 512
      responses:
        "200":
          description: Terrain grid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerrainData"
        "404":
          description: Dataset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Upstream data fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /user/datasets/{id}/terrain:
    get:
      operationId: getUserDatasetsIdTerrain
      tags: [user-datasets]
      summary: Get full terrain grid for a saved user dataset
      security:
        - clerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Terrain grid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerrainData"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Dataset not found or does not belong to current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /routes:
    get:
      operationId: getRoutes
      tags: [routes]
      summary: List saved navigation routes for a dataset
      description: Returns all routes belonging to the authenticated user for the given dataset, ordered by creation time.
      parameters:
        - in: query
          name: datasetId
          required: true
          schema:
            type: string
          description: Dataset ID to filter routes by
      responses:
        "200":
          description: List of saved routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedRoute"
        "400":
          description: datasetId is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a new saved navigation route
      description: Saves a named waypoint sequence for the given dataset. Waypoints must contain at least 2 and at most 20 points.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [datasetId, name, waypoints, totalDistanceM]
              properties:
                datasetId:
                  type: string
                name:
                  type: string
                  maxLength: 120
                waypoints:
                  type: array
                  minItems: 2
                  maxItems: 20
                  items:
                    type: object
                    properties:
                      lon:
                        type: number
                      lat:
                        type: number
                      depth:
                        type: number
                totalDistanceM:
                  type: number
      responses:
        "201":
          description: Route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedRoute"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /routes/{id}:
    patch:
      operationId: patchRoute
      tags: [routes]
      summary: Rename a saved navigation route
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 120
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedRoute"
        "400":
          description: Invalid id or body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Route not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete a saved navigation route
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Route deleted
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Route not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /terrain/bundles:
    post:
      tags: [terrain]
      summary: Trigger on-demand bathymetry bundle download
      description: |
        Creates or re-queues a background job that fetches a full bathymetry
        grid for the specified preset and writes it to the authenticated user's
        private GCS storage.  Returns immediately with the job ID and status.
        Poll `GET /terrain/bundles/{presetId}/status` for progress.
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [presetId]
              properties:
                presetId:
                  type: string
                  description: Preset dataset ID to fetch bathymetry for
      responses:
        "200":
          description: Bundle already exists and is available
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [complete]
                  message:
                    type: string
        "202":
          description: Job created or already in progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending, running]
                  message:
                    type: string
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Unknown preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: Preset has no fetchStrategy configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /terrain/bundles/{presetId}/status:
    get:
      tags: [terrain]
      summary: Get bundle job status
      description: Returns the current state of the on-demand download job for this preset.
      security:
        - clerkAuth: []
      parameters:
        - in: path
          name: presetId
          required: true
          schema:
            type: string
          description: Preset dataset ID
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending, running, complete, error]
                  progressNote:
                    type: string
                    nullable: true
                  errorMessage:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                    format: date-time
                  completedAt:
                    type: string
                    format: date-time
                    nullable: true
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No job found for this preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /terrain/bundles/{presetId}:
    get:
      tags: [terrain]
      summary: Get processed bathymetry bundle
      description: |
        Returns the full processed bathymetry bundle for a preset if the
        download job is complete.  Returns 202 if still pending/running,
        or 404 if no job exists.
      security:
        - clerkAuth: []
      parameters:
        - in: path
          name: presetId
          required: true
          schema:
            type: string
          description: Preset dataset ID
      responses:
        "200":
          description: Complete bathymetry bundle
          content:
            application/json:
              schema:
                type: object
                description: Processed depth grid bundle with metadata
        "202":
          description: Download still in progress or job errored
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending, running, error]
                  message:
                    type: string
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No bundle found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
        details:
          type: string
          description: Human-readable description of what went wrong
        requestId:
          type: string
          description: Optional tracing ID for correlating this error with server logs
      required:
        - error
        - details
...
    TerrainData:
      type: object
      properties:
        datasetId:
          type: string
        name:
          type: string
        waterType:
          type: string
          enum: [saltwater, freshwater]
        resolution:
          type: integer
          description: Grid side length N (grid is NxN)
        width:
          type: integer
        height:
          type: integer
        depths:
          type: array
          items:
            type: number
            nullable: true
          description: Row-major flat array of depth values (metres, positive = below surface). Null entries represent unfilled cells (survey gaps) that the IDW interpolation could not reach; clients should render them as flat no-data tiles at the water surface.
        minDepth:
          type: number
        maxDepth:
          type: number
        minLon:
          type: number
        maxLon:
          type: number
        minLat:
          type: number
        maxLat:
          type: number
        centerLon:
          type: number
        centerLat:
          type: number
        topography:
          type: array
          items:
            type: number
          description: Row-major flat array of above-water elevation values (metres above sea level, 0 for water cells). Same NxN shape as `depths`. Omitted when the dataset has no above-water terrain.
        hasTopography:
          type: boolean
          description: True when this terrain grid includes a non-empty `topography` array. Clients can use this to enable/hide the landmass visualisation.
        synthetic:
          type: boolean
          description: "Deprecated: use dataSource instead. True when the grid was produced from the synthetic fbm fallback."
        dataSource:
          type: string
          enum: [ncei, gebco, synthetic, twdb, usace, usgs-3dep]
          description: |
            Which upstream data service produced this grid (bathymetry primary).
            ncei      — NCEI Bag Mosaic WCS (high-resolution multibeam survey)
            gebco     — GEBCO 2024 WCS (~400 m global grid)
            synthetic — fbm fallback used when all upstream services are unreachable
            twdb      — TWDB Reservoir Volumetric & Sedimentation Survey
            usace     — USACE hydrographic survey
            usgs-3dep — USGS 3DEP best-available DEM (pre-impoundment basin + topography)
        bathymetrySource:
          type: string
          enum: [ncei, gebco, synthetic, twdb, usace, usgs-3dep]
          description: Per-layer provenance for the below-water (bathymetry) layer when it has a different source than topography (e.g. inland reservoirs).
        topographySource:
          type: string
          enum: [ncei, gebco, synthetic, twdb, usace, usgs-3dep]
          description: Per-layer provenance for the above-water (topography) layer when it has a different source than bathymetry.
        bathymetrySourceLabel:
          type: string
          description: Display label for the bathymetry source (overrides the default per-source label in the UI).
        topographySourceLabel:
          type: string
          description: Display label for the topography source (overrides the default per-source label in the UI).
        bathymetryCreditUrl:
          type: string
          description: Credit URL surfaced next to the bathymetry source badge.
        topographyCreditUrl:
          type: string
          description: Credit URL surfaced next to the topography source badge.
        habitatPolygons:
          type: object
          description: |
            EFH (Essential Fish Habitat) FeatureCollection embedded directly in the terrain
            response for user-saved noaa-efh-* catalog datasets. Present only when the
            dataset carries polygon habitat data (i.e. was saved from a NOAA EFH catalog
            entry). Clients should render these polygons instead of (or in addition to)
            fetching from /efh when this field is present.
          properties:
            type:
              type: string
              enum: [FeatureCollection]
            features:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    enum: [Feature]
                  properties:
                    $ref: "#/components/schemas/EfhSpeciesProperties"
                  geometry:
                    type: object
                    additionalProperties: true
                required: [type, properties, geometry]
            metadata:
              type: object
              additionalProperties: true
          required: [type, features]
      required:
        - datasetId
        - name
        - waterType
        - resolution
        - width
        - height
        - depths
        - minDepth
        - maxDepth
        - minLon
        - maxLon
        - minLat
        - maxLat
        - centerLon
        - centerLat
...
    EfhSpeciesProperties:
      type: object
      properties:
        species:
          type: string
          description: Scientific species name (snake_case)
        commonName:
          type: string
        fmp:
          type: string
          description: Fishery Management Plan name
        depthRangeM:
          type: array
          items:
            type: number
          description: "[minDepth, maxDepth] in metres"
        habitatDescription:
          type: string
        lifeStage:
          type: string
          description: Life stages covered by this Essential Fish Habitat polygon (e.g. "Juveniles & Adults")
        season:
          type: string
          description: Seasonality / temporal window for this Essential Fish Habitat designation (e.g. "Year-round", "Spawning Feb–Apr")
        source:
          type: string
        creditUrl:
          type: string
        color:
          type: string
          description: Suggested hex color for rendering
      required:
        - species
        - commonName
        - fmp
        - depthRangeM
        - habitatDescription
        - source
        - creditUrl
        - color
...
    SavedRoute:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        datasetId:
          type: string
        name:
          type: string
        waypoints:
          type: array
          items:
            type: object
            properties:
              lon:
                type: number
              lat:
                type: number
              depth:
                type: number
        waypointCount:
          type: integer
        totalDistanceM:
          type: number
        createdAt:
          type: string
      required:
        - id
        - userId
        - datasetId
        - name
        - waypoints
        - waypointCount
        - totalDistanceM
...